# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.Guild:
    properties:
      allowModeration:
        type: boolean
      banTime:
        type: integer
      disabledCommands:
        type: string
      guildID:
        type: string
      guildName:
        type: string
      lvlChannel:
        type: string
      lvlReplace:
        type: boolean
      lvlResponse:
        type: integer
      maxWarns:
        type: integer
      prefix:
        type: string
      privateWelcomeMsg:
        type: string
      reportChannel:
        type: string
      welcomeChannel:
        type: string
      welcomeMsg:
        type: string
    type: object
  models.Member:
    properties:
      guildID:
        type: string
      joinedAt:
        format: date-time
        type: string
      left:
        type: integer
      level:
        type: integer
      memberID:
        type: string
      xp:
        type: integer
    type: object
...
host: cardinal.gyroskan.com
info:
  contact:
    email: gyroskan@gmail.com
    name: API Support
  description: The API to interact with cardinal discord bot database.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Cardinal API
  version: "0.1"
paths:
  /guilds/:
    get:
      description: Fetch all guilds.
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Guild'
            type: array
        "403":
          description: Forbidden
        "500":
          description: Server error
      summary: Get all Guilds
      tags:
      - Guilds
    post:
      description: Creates a new Guild with the given values
      parameters:
      - description: Provide the guild values
        in: body
        name: guild
        required: true
        schema:
          $ref: '#/definitions/models.Guild'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Guild'
        "400":
          description: Bad Request
        "403":
          description: Forbidden
        "409":
          description: Conflict
        "500":
          description: Server error
      summary: Create guild
      tags:
      - Guilds
  /guilds/{guildID}:
    delete:
      consumes:
      - application/json
      description: Delete a guild
      parameters:
      - description: Guild id
        in: path
        name: guildID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "206":
          description: No Content
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "500":
          description: Server Error
      summary: Delete guild
      tags:
      - Guilds
    get:
      description: Fetch a specific guild
      parameters:
      - description: guild id
        in: path
        name: guildID
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Guild'
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "500":
          description: Server error
      summary: Get one guild
      tags:
      - Guilds
    patch:
      consumes:
      - application/json
      description: Update fields of a guild
      parameters:
      - description: Guild id
        in: path
        name: guildID
        required: true
        type: string
      - description: Guild modifications
        in: body
        name: guild
        required: true
        schema:
          $ref: '#/definitions/models.Guild'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Guild'
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "500":
          description: Server Error
      summary: Update guild values
      tags:
      - Guilds
...
  /guilds/{guildID}/members:
    get:
      description: Fetch all members of the guild.
      parameters:
      - description: guild id
        in: path
        name: guildID
        required: true
        type: string
      - default: 1
        description: limit to fetch
        in: query
        name: limit
        type: integer
      - default: "0"
        description: higher last id fetched
        in: query
        name: after
        type: string
      responses:
        "200":
          description: OK
        "403":
          description: Forbidden
        "500":
          description: Server error
      summary: Get Guild Members
      tags:
      - Members
    post:
      consumes:
      - application/json
      description: Create a new member from a guild.
      parameters:
      - description: guild id
        in: path
        name: guildID
        required: true
        type: string
      - description: Member values
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.Member'
      produces:
      - application/json
      responses:
        "201":
          description: Created member
          schema:
            $ref: '#/definitions/models.Member'
        "400":
          description: Wrong values
        "403":
          description: Forbidden
        "500":
          description: Server Error
      summary: Create member
      tags:
      - Members
  /guilds/{guildID}/members/{memberID}:
    delete:
      consumes:
      - application/json
      description: Delete a guild member
      parameters:
      - description: Guild id
        in: path
        name: guildID
        required: true
        type: string
      - description: Member id
        in: path
        name: memberID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "206":
          description: No Content
        "403":
          description: Forbidden
        "404":
          description: Not Fountd
        "500":
          description: Server Error
      summary: Delete guild member
      tags:
      - Members
    get:
      description: Fetch the member of the guild.
      parameters:
      - description: guild id
        in: path
        name: guildID
        required: true
        type: string
      - description: member id
        in: path
        name: memberID
        required: true
        type: string
      responses:
        "200":
          description: OK
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "500":
          description: Server error
      summary: Get one Guild Member
      tags:
      - Members
    patch:
      consumes:
      - application/json
      description: Update fields of a guild's member
      parameters:
      - description: Guild id
        in: path
        name: guildID
        required: true
        type: string
      - description: Guild id
        in: path
        name: memberID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "403":
          description: Forbidden
        "404":
          description: Not Fountd
        "500":
          description: Server Error
      summary: Update member value
      tags:
      - Members
...
  /guilds/{guildID}/members/{memberID}/reset:
    post:
      consumes:
      - application/json
      description: Reset level and xp for the specific guild's members.
      parameters:
      - description: Guild id
        in: path
        name: guildID
        required: true
        type: string
      - description: Guild id
        in: path
        name: memberID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Member reset
        "403":
          description: Forbidden
        "500":
          description: Server Error
      summary: Reset member from a guild
      tags:
      - Members
...
  /guilds/{guildID}/members/reset:
    post:
      consumes:
      - application/json
      description: Reset level and xp for all guild's members.
      parameters:
      - description: Guild id
        in: path
        name: guildID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Member reset
        "403":
          description: Forbidden
        "500":
          description: Server Error
      summary: Reset guild's members
      tags:
      - Members
  /guilds/{guildID}/reset:
    post:
      consumes:
      - application/json
      description: |-
        Reset guild parameters to default values.
        Do not change members values.
      parameters:
      - description: Guild id
        in: path
        name: guildID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Guild'
        "403":
          description: Forbidden
        "500":
          description: Server Error
      summary: Reset guild
      tags:
      - Guilds
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"